refactor: Replace basic WebSocket with enhanced implementation
- Remove old basic WebSocket transport implementation
- Rename WebSocketEnhancedTransport to WebSocketServerTransport
- Remove WEBSOCKET_ENHANCED from TransportType enum
- Update all references to use single 'websocket' transport type
- Update tests and documentation accordingly
- Remove old skipped test file

The enhanced WebSocket implementation is now the default and only
WebSocket transport, providing production-ready features like
connection health monitoring, message queuing, and graceful shutdown.
