+++ ai_whisperer/mcp/server/runner.py
@@ -122,7 +122,7 @@ def main():
     )
     parser.add_argument(
         '--transport',
-        choices=['stdio', 'websocket', 'websocket_enhanced', 'sse'],
+        choices=['stdio', 'websocket', 'sse'],
         default='stdio',
         help='Transport type (default: stdio)'
     )
